requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
pandas==2.0.3
